week2_function_agents/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.1.0
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
readability-lxml==0.8.4.1
requests==2.32.5
rich==14.1.0
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
